clang-tidy: config decisions + fix concurrency/bitwise/macro-to-enum
.clang-tidy: exclude two more families with rationale (DeprecatedOrUnsafeBufferHandling
— Annex K _s funcs unavailable on macOS/glibc/musl; llvm-header-guard — emits
path-dependent guards, existing REMEMBER_<FILE>_H are correct). Add an idiom
allowlist for readability-identifier-length so it stays ON for genuinely cryptic
names while accepting the established one/two-letter idioms (matches .swiftlint.yml).

Fixes:
- concurrency-mt-unsafe: gmtime -> gmtime_r (3 sites, real reentrancy); getenv (2)
  and strerror (1) NOLINT'd with reason (no portable thread-safe variant;
  single-threaded CLI).
- hicpp-signed-bitwise: unsigned operands in normalize.c; SQLITE_OPEN_* or NOLINT'd
  (sqlite's own signed flag macros).
- cppcoreguidelines-avoid-non-const-global-variables: NOLINT the deliberate test
  seams (appio g_out/g_err settable streams; store_sqlite fault-injection hooks).
- macro-to-enum: convert integral constant macros to enums.

+++ src/normalize.c
@@ -99,7 +99,7 @@ static bool utf8_cont(const char *s, size_t len, size_t i, size_t need, unsigned
         if ((cc & 0xC0U) != 0x80U) {
             return false;
         }
-        *cp = (*cp << 6) | (cc & 0x3FU);
+        *cp = (*cp << 6U) | (cc & 0x3FU);
     }
     return true;
 }
@@ -296,7 +296,7 @@ void body_hash_hex(const void *data, size_t len, char out_hex[REMEMBER_SHA256_HE
         unsigned char b = digest[i];
         size_t hi = i * 2U;
         size_t lo = hi + 1U;
-        out_hex[hi] = k_hex[(b >> 4) & 0x0FU];
+        out_hex[hi] = k_hex[((unsigned int)b >> 4U) & 0x0FU];
         out_hex[lo] = k_hex[b & 0x0FU];
     }
     out_hex[REMEMBER_SHA256_HEX_LEN] = '\0';
